skydome values precission reduced to Float16

+++ src/gl/Shader.ts
@@ -205,6 +205,7 @@ export class Shader {
             }
 
             if (name.toString() === "skydome") {
+                // console.log('uniforms.skydomeTextureSize: ', uniforms.skydomeTextureSize);
                 let rgbList = new Float32Array(uniforms.skydomeTextureSize * uniforms.skydomeTextureSize * 3);
                 for (let i = 0; i < uniforms.skydomeWidth * uniforms.skydomeHeight; i++) {
                     rgbList[i * 3 + 0] = uniforms.skydome.data[i * 4 + 0];
@@ -220,7 +221,7 @@ export class Shader {
                 gl.texParameterf(gl.TEXTURE_2D, gl.TEXTURE_WRAP_S, gl.CLAMP_TO_EDGE);
                 gl.texParameterf(gl.TEXTURE_2D, gl.TEXTURE_WRAP_T, gl.CLAMP_TO_EDGE);
 
-                gl.texImage2D(gl.TEXTURE_2D, 0, gl.RGB32F, uniforms.skydomeTextureSize, uniforms.skydomeTextureSize, 0, gl.RGB, gl.FLOAT, rgbList);
+                gl.texImage2D(gl.TEXTURE_2D, 0, gl.RGB16F, uniforms.skydomeTextureSize, uniforms.skydomeTextureSize, 0, gl.RGB, gl.FLOAT, rgbList);
 
                 let skydomeTextureLocation = gl.getUniformLocation(this._program, "skydomeTexture");
                 gl.uniform1i(skydomeTextureLocation, 6);
